task: make dedicated component for api endpoints
