+++ src/apps/content-editor/src/app/views/ItemEdit/components/ItemEditHeader/MoreMenu.tsx
@@ -17,8 +17,6 @@ import {
   CodeRounded,
   DeleteRounded,
   CheckRounded,
-  DesignServicesRounded,
-  VisibilityRounded,
   KeyboardArrowRightRounded,
 } from "@mui/icons-material";
 import { useState } from "react";
@@ -35,6 +33,7 @@ import { DeleteItemDialog } from "./DeleteItemDialog";
 import { useGetContentModelsQuery } from "../../../../../../../../shell/services/instance";
 import { usePermission } from "../../../../../../../../shell/hooks/use-permissions";
 import { CascadingMenuItem } from "../../../../../../../../shell/components/CascadingMenuItem";
+import { APIEndpoints } from "../../../../components/APIEndpoints";
 
 export const MoreMenu = () => {
   const { modelZUID, itemZUID } = useParams<{
@@ -44,11 +43,7 @@ export const MoreMenu = () => {
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const [isCopied, setIsCopied] = useState(false);
   const [showDuplicateItemDialog, setShowDuplicateItemDialog] = useState(false);
-  const [showApiEndpoints, setShowApiEndpoints] = useState<null | HTMLElement>(
-    null
-  );
   const [showDeleteItemDialog, setShowDeleteItemDialog] = useState(false);
-  const [apiEndpointType, setApiEndpointType] = useState("quick-access");
   const history = useHistory();
   const codePath = useFilePath(modelZUID);
   const { data: contentModels } = useGetContentModelsQuery();
@@ -197,79 +192,5 @@ export const MoreMenu = () => {
         <DeleteItemDialog onClose={() => setShowDeleteItemDialog(false)} />
       )}
     </>
-  );
-};
-
-type APIEndpointsProps = {
-  type: Extract<ApiType, "quick-access" | "site-generators">;
-};
-const APIEndpoints = ({ type }: APIEndpointsProps) => {
-  const { itemZUID } = useParams<{
-    itemZUID: string;
-  }>();
-  const item = useSelector(
-    (state: AppState) => state.content[itemZUID] as ContentItem
-  );
-  const instance = useSelector((state: AppState) => state.instance);
-  const { data: domains } = useGetDomainsQuery();
-
-  const apiTypeEndpointMap: Partial<Record<ApiType, string>> = {
-    "quick-access": `/-/instant/${itemZUID}.json`,
-    "site-generators": item ? `/${item?.web?.path}/?toJSON` : "/?toJSON",
-  };
-
-  const liveDomain = domains?.find((domain) => domain.branch == "live");
-
-  return (
-    <MenuList>
-      <MenuItem
-        onClick={() => {
-          window.open(
-            // @ts-expect-error config not typed
-            `${CONFIG.URL_PREVIEW_PROTOCOL}${instance.randomHashID}${CONFIG.URL_PREVIEW}${apiTypeEndpointMap[type]}`,
-            "_blank"
-          );
-        }}
-      >
-        <ListItemIcon>
-          <DesignServicesRounded />
-        </ListItemIcon>
-        <Typography
-          variant="inherit"
-          noWrap
-          sx={{
-            width: 172,
-          }}
-        >
-          {/* @ts-expect-error config not typed */}
-          {`${instance.randomHashID}${CONFIG.URL_PREVIEW}${apiTypeEndpointMap[type]}`}
-        </Typography>
-        <Chip size="small" label="Dev" />
-      </MenuItem>
-      {liveDomain && (
-        <MenuItem
-          onClick={() => {
-            window.open(
-              `https://${liveDomain.domain}${apiTypeEndpointMap[type]}`,
-              "_blank"
-            );
-          }}
-        >
-          <ListItemIcon>
-            <VisibilityRounded />
-          </ListItemIcon>
-          <Typography
-            variant="inherit"
-            noWrap
-            sx={{
-              width: 172,
-            }}
-          >
-            {`${liveDomain.domain}${apiTypeEndpointMap[type]}`}
-          </Typography>
-          <Chip size="small" label="Prod" />
-        </MenuItem>
-      )}
-    </MenuList>
   );
 };
